editor: Include frame in editor.
Makes it easier to render frame semantically as one big piece,
as well as to divide it into parts if it doest fit onto a 3D printer's
print bed.

This commit includes a "Hack to load older format that didn't model the frame"
in the editor for converting my existing field JSON files.
It will be removed in the next commit.